added avoind mutable globals analyzer
