+++ cmd/avoid-mutable-globals/example.go
@@ -1,0 +1,11 @@
+package main
+
+var _someData = 123
+
+func sign(msg string) string {
+	if _someData > 10 {
+		return msg + "123"
+	}
+
+	return msg
+}
