--- cmd/go-printf-func-name/main.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"golang-linter/pkg/analyzer"
-	"golang.org/x/tools/go/analysis/singlechecker"
-)
-
-func main() {
-	singlechecker.Main(analyzer.Analyzer)
-}
